# sql/catalog: reduce allocations when checking for hydratable types