--- a/pkg/sql/catalog/typedesc/table_implicit_record_type.go
+++ b/pkg/sql/catalog/typedesc/table_implicit_record_type.go
@@ -187,6 +187,9 @@ func (v *tableImplicitRecordType) ForEachUDTDependentForHydration(_ func(t *type
 	return nil
 }
 
+// MaybeRequiresTypeHydration implements the catalog.Descriptor interface.
+func (v *tableImplicitRecordType) MaybeRequiresTypeHydration() bool { return false }
+
 // TypeDesc implements the catalog.TypeDescriptor interface.
 func (v *tableImplicitRecordType) TypeDesc() *descpb.TypeDescriptor {
 	v.panicNotSupported("TypeDesc")