The `catalog.MaybeRequiresHydration` function has been replaced with the
`Descriptor.MaybeRequiresTypeHydration` interface method. This avoids a
heap allocation that was required in `catalog.MaybeRequiresHydration`.
Arguments to interface methods typically escape to the heap, so the
closure passed to `Descriptor.ForEachUDTDependentForHydration` was heap
allocated. The new interface method avoids this.

Epic: None

Release note: None