@mgartner 2024-10-10
`BenchmarkEndToEnd`, which essentially starts a server and repeatedly runs queries against it, shows the improvement. I just ran subset of the queries—the simplest ones will show the most improvement in this case.

```
./dev bench pkg/sql/opt/bench -f='BenchmarkEndToEnd/kv-read$/' --bench-mem -v --count 5 --ignore-cache
```

```
name                                     old time/op    new time/op    delta
EndToEnd/kv-read/vectorize=on/Simple        158µs ± 2%     168µs ± 8%    ~     (p=0.151 n=5+5)
EndToEnd/kv-read/vectorize=on/Prepared     97.8µs ± 2%   101.3µs ± 6%    ~     (p=0.222 n=5+5)
EndToEnd/kv-read/vectorize=off/Simple       166µs ± 4%     164µs ± 2%    ~     (p=0.841 n=5+5)
EndToEnd/kv-read/vectorize=off/Prepared     104µs ± 3%     103µs ±10%    ~     (p=0.421 n=5+5)

name                                     old alloc/op   new alloc/op   delta
EndToEnd/kv-read/vectorize=on/Simple       39.9kB ± 4%    39.2kB ± 1%    ~     (p=0.111 n=5+4)
EndToEnd/kv-read/vectorize=on/Prepared     27.7kB ± 1%    27.9kB ± 5%    ~     (p=1.000 n=4+5)
EndToEnd/kv-read/vectorize=off/Simple      40.9kB ± 0%    41.4kB ± 5%    ~     (p=0.730 n=4+5)
EndToEnd/kv-read/vectorize=off/Prepared    29.5kB ± 3%    29.3kB ± 4%    ~     (p=1.000 n=5+5)

name                                     old allocs/op  new allocs/op  delta
EndToEnd/kv-read/vectorize=on/Simple          413 ± 0%       389 ± 0%  -5.81%  (p=0.029 n=4+4)
EndToEnd/kv-read/vectorize=on/Prepared        298 ± 3%       286 ± 3%  -4.10%  (p=0.032 n=5+5)
EndToEnd/kv-read/vectorize=off/Simple         403 ± 0%       378 ± 0%  -6.14%  (p=0.029 n=4+4)
EndToEnd/kv-read/vectorize=off/Prepared       287 ± 3%       275 ± 6%    ~     (p=0.087 n=5+5)
```